restructured

+++ main.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Created by JetBrains PhpStorm.
  * User: ewdchn
@@ -6,74 +7,90 @@
  * Time: 5:37 PM
  * To change this template use File | Settings | File Templates.
  */
-
-require_once 'Crawler_L0.php';
-require_once 'Parser_L0.php';
+require_once 'Crawler.php';
+require_once 'Parser.php';
 require_once 'test.php';
 
-if (strtoupper(substr(PHP_OS, 0, 3)) === 'WIN')
-{ echo "platform:windows\n";define("DIR_SEP",'\\');} else {define ("DIR_SEP",'\/');}
+if (strtoupper(substr(PHP_OS, 0, 3)) === 'WIN') {
+    echo "platform:windows\n";
+    define("DIR_SEP", '\\');
+} else {
+    define("DIR_SEP", '\/');
+}
+define('resultCnt', 522);
+define('TmpDir', "tmp");
 
-CONST resultCnt = 522;
-CONST L0_TmpDir = "L0";
-$level0eids=array();
-$level0 = array();
-$crawler = NULL;
+$L0eids = array();
+$L0Data = array();
 
-function searchCrawl(){
+function searchGrab() {
     echo "Start Crawling L0...\n";
 
     \Common\checkOrCreateDir(L0_TmpDir);
 
-    foreach (range(1,11) as $page){
-        $fileName = L0_TmpDir.DIR_SEP.'page'.$page.'.html';
-        file_put_contents($fileName,\Crawler\Crawler::handleSearch($page));
+    foreach (range(1, 11) as $page) {
+        $fileName = L0_TmpDir . DIR_SEP . 'page' . $page . '.html';
+        file_put_contents($fileName, \Crawler\Crawler::handleSearch($page));
     }
     echo "done\n";
 }
 
-function searchParse(&$_eidArray){
+function searchParse(&$_eidArray) {
     //parse each page of search results and output the eid of result entries as array
-    foreach(range(1,1) as $page){
-        $fileName = L0_TmpDir.DIR_SEP.'page'.$page.'.html';
+    foreach (range(1, 11) as $page) {
+        $fileName = L0_TmpDir . DIR_SEP . 'page' . $page . '.html';
 //        echo "parsing ",$fileName,"\n";
-        $result = \Parser_L0\parsePage($fileName);
+        $result = \Parser\parseSearchResultPage($fileName);
 //        echo ",",count($result),",";
-        $_eidArray =  array_merge($_eidArray,$result);
+        $_eidArray = array_merge($_eidArray, $result);
     }
 }
 
-function L0GrabEntries($_eidArray,&$_L0){
+function L0Grab($_eidArray) {
     echo "Start Grabbing L0 Entries...";
-    foreach($_eidArray as $key=>$eid){
+    foreach ($_eidArray as $key => $eid) {
+
         echo ".";
-
         $response = \Crawler\Crawler::grabPaperEntry($eid);
-        $_L0[$key] = \Parser\parseEntry($response);
 
-        $fileName =  L0_TmpDir.DIR_SEP.'entry_'.$key.'.html';
-        file_put_contents($fileName,$response);
+        $fileName = L0_TmpDir . DIR_SEP . 'entry_' . $eid . '.html';
+        if (!file_exists($fileName)) {
+            file_put_contents($fileName, $response);
+        }
     }
     echo "\ndone\n";
 }
 
-
-
-
-
+function L0Parse($_eidArray, &$_L0) {
+    if (empty($_eidArray)) {
+        throw new \Exception("array is empty");
+    }
+    foreach ($_eidArray as $key => $eid) {
+        $fileName = L0_TmpDir . DIR_SEP . 'entry_' . $eid . '.html';
+        $response = file_get_contents($fileName);
+        $_L0[$key] = \Parser\parseEntry($response, $eid);
+    }
+}
 
 set_time_limit(0);
-/*------------------------execution starts here-----------------------*/
+/* ------------------------execution starts here----------------------- */
 
 //// crawl level 0
-searchCrawl();
-searchParse($level0eids);
+//searchGrab();
+//searchParse($L0eids);
+//file_put_contents("l0eid.txt", serialize($L0eids));
 //return;
-
-
+//
+//
+//
+//
 //parse/extract level 0
-echo count($level0eids)."\n";
-L0GrabEntries($level0eids,$level0);
+$L0eids = unserialize(file_get_contents("l0eid.txt"));
+//echo count($L0eids) . "\n";
+//L0Grab($L0eids);
+//return;
+L0Parse($L0eids);
+
 return;
 
 //echo "\n",count($level0eids);
@@ -84,5 +101,4 @@ return;
 
 
 return 0;
-
 ?>
